requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
twilio==8.10.0
python-dotenv==1.0.0
google-genai==1.0.0